requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
asttokens
attrs
backoff
blinker
cachetools
certifi
charset-normalizer
click
colorama
comm
dataclasses-json
debugpy
decorator
executing
fastapi
filetype
frozenlist
gitdb
gitpython
google-ai-generativelanguage
google-api-core
google-auth
google-genai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httptools
httpx
httpx-sse
idna
ipykernel
ipython
ipython-pygments-lexers
jedi
jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter-client
jupyter-core
langchain
langchain-community
langchain-core
langchain-experimental
langchain-google-genai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
markupsafe
marshmallow
matplotlib-inline
multidict
mypy-extensions
narwhals
nest-asyncio
numpy
orjson
ormsgpack
packaging
pandas
parso
pillow
platformdirs
prompt-toolkit
propcache
proto-plus
protobuf
psutil
pure-eval
pyarrow
pyasn1
pyasn1-modules
pydantic
pydantic-core
pydantic-settings
pydeck
pygments
python-dateutil
python-dotenv
pytz
pyyaml
pyzmq
referencing
requests
requests-toolbelt
rpds-py
rsa
six
smmap
sniffio
sqlalchemy
stack-data
starlette
streamlit
streamlit-chat
streamlit-searchbox
tenacity
toml
tornado
traitlets
typing-extensions
typing-inspect
typing-inspection
tzdata
urllib3
uvicorn
watchdog
watchfiles
wcwidth
websockets
xxhash
yarl
zstandard
